Refactored SearchableBucketList
