+++ src/main/QueryServer.cpp
@@ -3,8 +3,8 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "main/QueryServer.h"
-#include "bucket/BucketListSnapshot.h"
 #include "bucket/BucketSnapshotManager.h"
+#include "bucket/SearchableBucketList.h"
 #include "ledger/LedgerTxnImpl.h"
 #include "util/Logging.h"
 #include "util/XDRStream.h" // IWYU pragma: keep
